// src/enums.tsx
export enum requestStatus {
    success = 'success',
    error = 'error'
}

export enum requestLabel {
    success = 'Успешно',
    error = 'Изменения не сохранены'
}

export enum requestText {
    success = 'Изменения успешно сохранены',
    error = 'Потеря интернет соединения'
}

export enum jsEvent {
    mouseenter = 'mouseenter',
    mouseleave = 'mouseleave'
}